requirements.txt
Flask
python-dotenv
line-bot-sdk>=3.0.0
requests
pandas
matplotlib
mpl_finance
gunicorn